+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Adds the /verify and /unverify admin commands.
+
 ### Fixed
 
 - Better sanity checking user input on the /info command.